+++ test/functional/test-url-replacements.js
@@ -23,7 +23,7 @@ import {
   resetScheduledElementForTesting,
 } from '../../src/custom-element';
 import {cidServiceForDocForTesting,} from
-    '../../extensions/amp-analytics/0.1/cid-impl';
+    '../../src/cid-impl';
 import {installCryptoService} from '../../src/service/crypto-impl';
 import {installDocService} from '../../src/service/ampdoc-impl';
 import {installDocumentInfoServiceForDoc,} from
